requirements.txt
fastapi
uvicorn
python-multipart
rembg
pillow
onnxruntime
